backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
lxml==5.2.1
xmlschema==3.3.1
anthropic==0.26.0
python-dotenv==1.0.1
pydantic==2.7.1
